bpo 39632: Fix ctypes variadic function call convention
On armhf and for variadic functions (and contrary to non-variadic
functions), the VFP co-processor registers are not used for float
argument parameter passing. This specificity was completely
disregarded by ctypes which used `ffi_prep_cif` systematically
to prepare the parameter passing of a function while it should
use `ffi_prep_cif_var` for variadic functions instead.

As such variadic function call with float arguments through ctypes
was broken on armhf targets.

This change fixes and improves ctypes variadic function support.
Functions that take a variable number of arguments should now be
declared using an Ellipsis (...) as their last argument (just as
a prototype for a C function). ctypes variadic functions declared
this way automatically perform the default argument promotion.
Small integer types are promoted to c_int and c_float is promoted
to c_double.

For reference/see also:
- https://sourceware.org/ml/libffi-discuss/2016/msg00043.html
- https://bugs.python.org/issue17580
- https://github.com/spacepy/spacepy/issues/20
- https://en.cppreference.com/w/c/language/conversion#Default_argument_promotions